+++ src/EntityGraphQL/Compiler/GqlNodes/ExecutableGraphQLStatement.cs
@@ -182,8 +182,32 @@ namespace EntityGraphQL.Compiler
                     // the full selection is now on the anonymous type returned by the selection without fields. We don't know the type until now
                     var newContextType = Expression.Parameter(runningContext.GetType(), "ctx_no_srv");
 
+                    // core context data is fetched. Now fetch all the bulk resolvers
+                    var bulkData = new Dictionary<string, object>();
+                    if (compileContext.BulkResolvers?.Any() == true)
+                    {
+                        foreach (var bulkResolver in compileContext.BulkResolvers)
+                        {
+                            // rebuild list expression on new context
+                            var listExpression = replacer.Replace(bulkResolver.ListExpression, node.Field!.ResolveExpression!, newContextType);
+                            var newParam = Expression.Parameter(listExpression.Type.GetEnumerableOrArrayType()!, "bulkList");
+                            // replace the data selection expression with the new context
+                            var expReplacer = new ExpressionReplacer(bulkResolver.ExtractedFields, newParam, false, false);
+                            var selection = expReplacer.Replace(bulkResolver.DataSelection.Body);
+                            var selectionLambda = Expression.Lambda(selection, newParam);
+                            selection = ExpressionUtil.MakeCallOnEnumerable(nameof(Enumerable.Select), new Type[] { newParam.Type, selection.Type }, listExpression, selectionLambda);
+
+                            var bulkDataArgs = Expression.Lambda(selection, newContextType).Compile().DynamicInvoke(new[] { runningContext });
+                            var parameters = new List<ParameterExpression> { bulkResolver.FieldExpression.Parameters.First() };
+                            var allArgs = new List<object?> { bulkDataArgs };
+                            var bulkLoader = GraphQLHelper.InjectServices(serviceProvider!, compileContext.Services, allArgs, bulkResolver.FieldExpression.Body, parameters, replacer);
+                            var dataLoaded = Expression.Lambda(bulkLoader, parameters).Compile().DynamicInvoke(allArgs.ToArray())!;
+                            bulkData[bulkResolver.Name] = dataLoaded;
+                        }
+                    }
+
                     // new context
-                    compileContext = new();
+                    compileContext = new(bulkData);
 
                     // we now know the selection type without services and need to build the full select on that type
                     // need to rebuild the full query
@@ -230,6 +254,12 @@ namespace EntityGraphQL.Compiler
             {
                 var returnType = typeof(List<>).MakeGenericType(expression.Type.GetEnumerableOrArrayType()!);
                 expression = Expression.Call(typeof(EnumerableExtensions), nameof(EnumerableExtensions.ToListWithNullCheck), new[] { expression.Type.GetEnumerableOrArrayType()! }, expression, Expression.Constant(node.Field!.ReturnType.TypeNotNullable));
+            }
+
+            if (compileContext.BulkData != null)
+            {
+                parameters.Add(compileContext.BulkParameter!);
+                allArgs.Add(compileContext.BulkData);
             }
 
             if (options.BeforeExecuting != null)