+++ src/EntityGraphQL/Schema/FieldToResolve.cs
@@ -1,10 +1,67 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
+using EntityGraphQL.Compiler;
+using EntityGraphQL.Compiler.Util;
 
 namespace EntityGraphQL.Schema;
-public class FieldToResolveWithArgs<TContext, TParams> : Field
+
+/// <summary>
+/// Represents a field with arguments that has a resolve expression set. The generics allow compile time checking of the bulk resolver.
+/// </summary>
+/// <typeparam name="TContext"></typeparam>
+/// <typeparam name="TParams"></typeparam>
+public class FieldWithContextAndArgs<TContext, TParams> : Field
 {
-    public FieldToResolveWithArgs(ISchemaProvider schema, ISchemaType fromType, string name, string? description, TParams argTypes) : base(schema, fromType, name, null, description, argTypes, SchemaBuilder.MakeGraphQlType(schema, typeof(object), null), null)
+    public FieldWithContextAndArgs(ISchemaProvider schema, ISchemaType fromType, string name, string? description, TParams argTypes)
+        : base(schema, fromType, name, null, description, argTypes, SchemaBuilder.MakeGraphQlType(schema, typeof(object), null), null)
+    {
+    }
+
+    public Field ResolveBulk<TService, TKey, TResult>(Expression<Func<TContext, TKey>> dataSelector, Expression<Func<IEnumerable<TKey>, IEnumerable<TParams>, TService, IDictionary<TKey, TResult>>> fieldExpression)
+    {
+        var extractor = new ExpressionExtractor();
+        var keyParam = dataSelector.Parameters.First();
+        var fields = extractor.Extract(dataSelector, keyParam, false)?.Select(i => new GraphQLExtractedField(Schema, i.Key, i.Value, keyParam))!;
+        ExtractedFieldsFromServices!.AddRange(fields);
+        BulkResolver = new BulkFieldResolverWithArgs<TContext, TParams, TService, TKey, TResult>($"bulk_{Name}", fieldExpression, dataSelector, fields);
+        Services.Add(fieldExpression.Parameters[1]);
+        return this;
+    }
+}
+
+/// <summary>
+/// Represents a field that has a resolve expression set. The generics allow compile time checking of the bulk resolver.
+/// </summary>
+/// <typeparam name="TContext"></typeparam>
+public class FieldWithContext<TContext> : Field
+{
+    public FieldWithContext(ISchemaProvider schema, ISchemaType fromType, string name, string? description, object? argTypes)
+        : base(schema, fromType, name, null, description, argTypes, SchemaBuilder.MakeGraphQlType(schema, typeof(object), null), null)
+    {
+    }
+
+    public Field ResolveBulk<TService, TKey, TResult>(Expression<Func<TContext, TKey>> dataSelector, Expression<Func<IEnumerable<TKey>, TService, IDictionary<TKey, TResult>>> fieldExpression)
+    {
+        var extractor = new ExpressionExtractor();
+        var keyParam = dataSelector.Parameters.First();
+        var fields = extractor.Extract(dataSelector, keyParam, false)?.Select(i => new GraphQLExtractedField(Schema, i.Key, i.Value, keyParam))!;
+        ExtractedFieldsFromServices!.AddRange(fields);
+        BulkResolver = new BulkFieldResolver<TContext, TService, TKey, TResult>($"bulk_{Name}", fieldExpression, dataSelector, fields);
+        Services.Add(fieldExpression.Parameters[1]);
+        return this;
+    }
+}
+
+/// <summary>
+/// Represents a field with arguments that still needs it's resolve expression to be set. The generics allow compile time checking of the expression.
+/// </summary>
+/// <typeparam name="TContext"></typeparam>
+/// <typeparam name="TParams"></typeparam>
+public class FieldToResolveWithArgs<TContext, TParams> : FieldWithContextAndArgs<TContext, TParams>
+{
+    public FieldToResolveWithArgs(ISchemaProvider schema, ISchemaType fromType, string name, string? description, TParams argTypes) : base(schema, fromType, name, description, argTypes)
     {
     }
 
@@ -13,41 +70,45 @@ public class FieldToResolveWithArgs<TContext, TParams> : Field
         SetUpField(fieldExpression, true, true);
         return this;
     }
-    public Field ResolveWithService<TService>(Expression<Func<TContext, TParams, TService, object>> fieldExpression)
+    public FieldWithContextAndArgs<TContext, TParams> ResolveWithService<TService>(Expression<Func<TContext, TParams, TService, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
-        Services = new[] { fieldExpression.Parameters[2] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[2] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TParams, TService1, TService2, object>> fieldExpression)
+    public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TParams, TService1, TService2, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
-        Services = new[] { fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TParams, TService1, TService2, TService3, object>> fieldExpression)
+    public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TParams, TService1, TService2, TService3, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
-        Services = new[] { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, object>> fieldExpression)
+    public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
-        Services = new[] { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
+    public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
-        Services = new[] { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5], fieldExpression.Parameters[6] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5], fieldExpression.Parameters[6] };
         return this;
     }
 }
 
-public class FieldToResolve<TContext> : Field
+/// <summary>
+/// Represents a field that still needs it's resolve expression to be set. The generics allow compile time checking of the expression.
+/// </summary>
+/// <typeparam name="TContext"></typeparam>
+public class FieldToResolve<TContext> : FieldWithContext<TContext>
 {
-    public FieldToResolve(ISchemaProvider schema, ISchemaType fromType, string name, string? description, object? argTypes) : base(schema, fromType, name, null, description, argTypes, SchemaBuilder.MakeGraphQlType(schema, typeof(object), null), null)
+    public FieldToResolve(ISchemaProvider schema, ISchemaType fromType, string name, string? description, object? argTypes) : base(schema, fromType, name, description, argTypes)
     {
     }
 
@@ -57,34 +118,34 @@ public class FieldToResolve<TContext> : Field
         return this;
     }
 
-    public Field ResolveWithService<TService>(Expression<Func<TContext, TService, object>> fieldExpression)
+    public FieldWithContext<TContext> ResolveWithService<TService>(Expression<Func<TContext, TService, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
-        Services = new[] { fieldExpression.Parameters[1] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[1] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TService1, TService2, object>> fieldExpression)
+    public FieldWithContext<TContext> ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TService1, TService2, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
-        Services = new[] { fieldExpression.Parameters[1], fieldExpression.Parameters[2] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TService1, TService2, TService3, object>> fieldExpression)
+    public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TService1, TService2, TService3, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
-        Services = new[] { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TService1, TService2, TService3, TService4, object>> fieldExpression)
+    public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
-        Services = new[] { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
         return this;
     }
-    public Field ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
+    public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
-        Services = new[] { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5] };
+        Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5] };
         return this;
     }
 }